requirements.txt
Pillow
protobuf
Requests
streamlit
